refactor #12 QuestionService.delete() 수정
- 인자로 questionId와 user를 받아서 delete 할수 있도록 수정

+++ src/main/java/com/codessquad/qna/service/QuestionService.java
@@ -38,8 +38,11 @@ public class QuestionService {
     }
 
     @Transactional
-    public void delete(Question question) {
-        questionRepository.delete(question);
+    public void delete(long questionId, User user) {
+        Question question = questionRepository.findById(questionId).orElseThrow(IllegalArgumentException::new);
+        if (verifyQuestion(question,user)) {
+            questionRepository.delete(question);
+        }
     }
 
     public boolean verifyQuestion(Question question, User sessionedUser) {
